Pull UInt8/UInt16 extensions to separate files
