+++ Sources/Swift6502/Extensions/UInt16+Extensions.swift
@@ -8,4 +8,8 @@ extension UInt16 {
     var lowByte: UInt8 {
         UInt8(self & 0xFF)
     }
+
+    static func createWord(highByte: UInt8, lowByte: UInt8) -> UInt16 {
+        (highByte.asWord << 8) | lowByte.asWord
+    }
 }
