+++ Sources/Swift6502/Extensions/UInt8+Extensions.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension UInt8 {
+    var asWord: UInt16 {
+        UInt16(self)
+    }
+}
